overhaul style, use signed interger by default, get rid of the supportVectorCalssifier base on ipopt lib.

+++ src/tests/Scaler_unit.cc
@@ -10,15 +10,15 @@ TEST(Scaler, Standard_test)
 {
   using Scaler=Preprocessing::StandardScaler;
 
-  unsigned numberOfFeatures=10;
-  unsigned numberOfData=100;
+  long numberOfFeatures=10;
+  long numberOfData=100;
 
   MatrixXd data=MatrixXd::Random(numberOfData,numberOfFeatures);
 
   Scaler scaler;
   MatrixXd transformedData=scaler.fitTransform(data);
 
-  for (size_t featIndex=0; featIndex<numberOfFeatures; ++featIndex){
+  for (long featIndex=0; featIndex<numberOfFeatures; ++featIndex){
     double mean=transformedData.col(featIndex).mean();
     double stdDev=Utilities::computeStandardDeviation(transformedData.col(featIndex));
     EXPECT_NEAR(mean, 0.0, 1e-10);
